backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
python-dotenv==1.0.1
qrcode==7.4.2
twilio==9.3.0
stripe==10.5.0